backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
httpx
